+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/ArmoredStatue.java
@@ -139,7 +139,7 @@ public class ArmoredStatue extends Statue {
 	}
 
 	@Override
-	public int defenseSkill(Char enemy) {
+	public long defenseSkill(Char enemy) {
 		return Math.round(armor.evasionFactor(this, super.defenseSkill(enemy)));
 	}
