+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/npcs/Blacksmith.java
@@ -393,7 +393,7 @@ public class Blacksmith extends NPC {
 	}
 
 	@Override
-	public int defenseSkill( Char enemy ) {
+	public long defenseSkill(Char enemy ) {
 		return INFINITE_EVASION;
 	}
