+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/Gold.java
@@ -90,7 +90,7 @@ public class Gold extends Item {
 	
 	@Override
 	public Item random() {
-		quantity = Dungeon.IntRange( 30 + Dungeon.escalatingDepth() * 10, 60 + Dungeon.escalatingDepth() * 20 );
+		quantity = Dungeon.LongRange( 30 + Dungeon.escalatingDepth() * 10, 60 + Dungeon.escalatingDepth() * 20 );
 		return this;
 	}
