+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/treasurebags/TenguTreasureBag.java
@@ -51,7 +51,7 @@ public class TenguTreasureBag extends TreasureBag {
             items.add(new TengusMask());
             Dungeon.LimitedDrops.TOME_OF_MASTERY.drop();
         }
-        items.add(new Gold().quantity(Random.Int( 625 + Dungeon.escalatingDepth() * 96, 1100 + Dungeon.escalatingDepth() * 130 )));
+        items.add(new Gold().quantity(Random.Long( 625 + Dungeon.escalatingDepth() * 96, 1100 + Dungeon.escalatingDepth() * 130 )));
         items.add(new Shuriken().upgrade(1).quantity(Random.Int(2, 10)));
         if (Dungeon.cycle > 0){
             items.add(new TenguShuriken().random().quantity(Random.Int(1, 3)));
